% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Salinity values of saline lagoons.}
\format{
A data frame with 136 rows and 2 variables:
\describe{
  \item{sample_ID}{Sample internal code}
  \item{salinity}{Sample salinity value}
}
}
\source{
{Triadó-Margarit, X., Capitán, J.A., Menéndez-Serra, M. et al. A Randomized Trait Community Clustering approach to unveil consistent environmental thresholds in community assembly. ISME J 13, 2681–2689.2019.}
}
\usage{
metadata
}
\description{
A dataset containing salinity values of 136 lagoons on the Monegros desert area.
}
\keyword{datasets}
