% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_presence_absence.R
\docType{data}
\name{table_presence_absence}
\alias{table_presence_absence}
\title{Genome presence-absence data of 136 saline lagoons.}
\format{
A data frame with 544 rows and 137 variables:
\describe{
  \item{genome}{Genome IMG code}
  \item{MON_10}{Sample presence-absence observations}
  \item{MON_100}{Sample presence-absence observations}
  \item{MON_101}{Sample presence-absence observations}
  \item{MON_103}{Sample presence-absence observations}
  \item{MON_104}{Sample presence-absence observations}
  \item{MON_106}{Sample presence-absence observations}
  \item{MON_107}{Sample presence-absence observations}
  \item{MON_108}{Sample presence-absence observations}
  \item{MON_109}{Sample presence-absence observations}
  \item{MON_11}{Sample presence-absence observations}
  \item{MON_110}{Sample presence-absence observations}
  \item{MON_111}{Sample presence-absence observations}
  \item{MON_112}{Sample presence-absence observations}
  \item{MON_113}{Sample presence-absence observations}
  \item{MON_114}{Sample presence-absence observations}
  \item{MON_116}{Sample presence-absence observations}
  \item{MON_117}{Sample presence-absence observations}
  \item{MON_118}{Sample presence-absence observations}
  \item{MON_119}{Sample presence-absence observations}
  \item{MON_12}{Sample presence-absence observations}
  \item{MON_120}{Sample presence-absence observations}
  \item{MON_122}{Sample presence-absence observations}
  \item{MON_123}{Sample presence-absence observations}
  \item{MON_124}{Sample presence-absence observations}
  \item{MON_125}{Sample presence-absence observations}
  \item{MON_126}{Sample presence-absence observations}
  \item{MON_127}{Sample presence-absence observations}
  \item{MON_128}{Sample presence-absence observations}
  \item{MON_129}{Sample presence-absence observations}
  \item{MON_13}{Sample presence-absence observations}
  \item{MON_130}{Sample presence-absence observations}
  \item{MON_131}{Sample presence-absence observations}
  \item{MON_133}{Sample presence-absence observations}
  \item{MON_134}{Sample presence-absence observations}
  \item{MON_135}{Sample presence-absence observations}
  \item{MON_136}{Sample presence-absence observations}
  \item{MON_137}{Sample presence-absence observations}
  \item{MON_138}{Sample presence-absence observations}
  \item{MON_139}{Sample presence-absence observations}
  \item{MON_14}{Sample presence-absence observations}
  \item{MON_140}{Sample presence-absence observations}
  \item{MON_141}{Sample presence-absence observations}
  \item{MON_142}{Sample presence-absence observations}
  \item{MON_144}{Sample presence-absence observations}
  \item{MON_145}{Sample presence-absence observations}
  \item{MON_146}{Sample presence-absence observations}
  \item{MON_147}{Sample presence-absence observations}
  \item{MON_148}{Sample presence-absence observations}
  \item{MON_15}{Sample presence-absence observations}
  \item{MON_17}{Sample presence-absence observations}
  \item{MON_18}{Sample presence-absence observations}
  \item{MON_19}{Sample presence-absence observations}
  \item{MON_2}{Sample presence-absence observations}
  \item{MON_20}{Sample presence-absence observations}
  \item{MON_21}{Sample presence-absence observations}
  \item{MON_22}{Sample presence-absence observations}
  \item{MON_23}{Sample presence-absence observations}
  \item{MON_24}{Sample presence-absence observations}
  \item{MON_25}{Sample presence-absence observations}
  \item{MON_26}{Sample presence-absence observations}
  \item{MON_27}{Sample presence-absence observations}
  \item{MON_28}{Sample presence-absence observations}
  \item{MON_29}{Sample presence-absence observations}
  \item{MON_30}{Sample presence-absence observations}
  \item{MON_31}{Sample presence-absence observations}
  \item{MON_32}{Sample presence-absence observations}
  \item{MON_33}{Sample presence-absence observations}
  \item{MON_34}{Sample presence-absence observations}
  \item{MON_35}{Sample presence-absence observations}
  \item{MON_36}{Sample presence-absence observations}
  \item{MON_37}{Sample presence-absence observations}
  \item{MON_38}{Sample presence-absence observations}
  \item{MON_39}{Sample presence-absence observations}
  \item{MON_4}{Sample presence-absence observations}
  \item{MON_40}{Sample presence-absence observations}
  \item{MON_41}{Sample presence-absence observations}
  \item{MON_42}{Sample presence-absence observations}
  \item{MON_43}{Sample presence-absence observations}
  \item{MON_44}{Sample presence-absence observations}
  \item{MON_45}{Sample presence-absence observations}
  \item{MON_46}{Sample presence-absence observations}
  \item{MON_47}{Sample presence-absence observations}
  \item{MON_48}{Sample presence-absence observations}
  \item{MON_49}{Sample presence-absence observations}
  \item{MON_5}{Sample presence-absence observations}
  \item{MON_50}{Sample presence-absence observations}
  \item{MON_51}{Sample presence-absence observations}
  \item{MON_52}{Sample presence-absence observations}
  \item{MON_53}{Sample presence-absence observations}
  \item{MON_54}{Sample presence-absence observations}
  \item{MON_55}{Sample presence-absence observations}
  \item{MON_56}{Sample presence-absence observations}
  \item{MON_57}{Sample presence-absence observations}
  \item{MON_58}{Sample presence-absence observations}
  \item{MON_59}{Sample presence-absence observations}
  \item{MON_60}{Sample presence-absence observations}
  \item{MON_61}{Sample presence-absence observations}
  \item{MON_62}{Sample presence-absence observations}
  \item{MON_63}{Sample presence-absence observations}
  \item{MON_64}{Sample presence-absence observations}
  \item{MON_65}{Sample presence-absence observations}
  \item{MON_66}{Sample presence-absence observations}
  \item{MON_67}{Sample presence-absence observations}
  \item{MON_68}{Sample presence-absence observations}
  \item{MON_69}{Sample presence-absence observations}
  \item{MON_7}{Sample presence-absence observations}
  \item{MON_70}{Sample presence-absence observations}
  \item{MON_71}{Sample presence-absence observations}
  \item{MON_72}{Sample presence-absence observations}
  \item{MON_73}{Sample presence-absence observations}
  \item{MON_74}{Sample presence-absence observations}
  \item{MON_75}{Sample presence-absence observations}
  \item{MON_76}{Sample presence-absence observations}
  \item{MON_77}{Sample presence-absence observations}
  \item{MON_78}{Sample presence-absence observations}
  \item{MON_79}{Sample presence-absence observations}
  \item{MON_8}{Sample presence-absence observations}
  \item{MON_80}{Sample presence-absence observations}
  \item{MON_81}{Sample presence-absence observations}
  \item{MON_82}{Sample presence-absence observations}
  \item{MON_83}{Sample presence-absence observations}
  \item{MON_84}{Sample presence-absence observations}
  \item{MON_85}{Sample presence-absence observations}
  \item{MON_86}{Sample presence-absence observations}
  \item{MON_88}{Sample presence-absence observations}
  \item{MON_9}{Sample presence-absence observations}
  \item{MON_90}{Sample presence-absence observations}
  \item{MON_91}{Sample presence-absence observations}
  \item{MON_92}{Sample presence-absence observations}
  \item{MON_93}{Sample presence-absence observations}
  \item{MON_94}{Sample presence-absence observations}
  \item{MON_95}{Sample presence-absence observations}
  \item{MON_96}{Sample presence-absence observations}
  \item{MON_97}{Sample presence-absence observations}
  \item{MON_98}{Sample presence-absence observations}
  \item{MON_99}{Sample presence-absence observations}
  ...
}
}
\source{
{Triadó-Margarit, X., Capitán, J.A., Menéndez-Serra, M. et al. A Randomized Trait Community Clustering approach to unveil consistent environmental thresholds in community assembly. ISME J 13, 2681–2689 (2019).}
}
\usage{
table_presence_absence
}
\description{
A dataset containing presence-absence data of 544 genomes on 136 saline lagoons of the Monegros desert area.
}
\keyword{datasets}
